backup/requirements.txt
PyMuPDF>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.1
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
tiktoken>=0.5.0
azure-monitor-opentelemetry>=1.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
